zookeeper: Optimize scraping time
These changes optimize scraping time from ~3m to ~8s for the same data.
We were using `FetchAllOffsets` for collecting the metrics, this method
althugh quite handy executes a request against zookeeper for each
topic/partition _sequentially_ resulting in long executing times for the
offshore DCs. We've migrated to use `FetchOffset` instead and
paralellize at the topic level, each topic is processed in a different
goroutine.

The `groupFilter` and `topicFilter` were ignored and we were collecting
all the topics/consumergroups registered in Zookeeper.