zookeeper: Filter topics not consumed by the group
The `FetchOffset` function wraps the `ErrNoNode` error and returns
`-1, nil` instead. This caused that a given topic not consumed by the
current consumer group was still reported as consumed, with a very high
lag.

The high value of the lag was the result of `-1 - currentOffset`.

This change also simplifies the groupFilter matching on the
consumergroup name.